- Rename BoxAction to BoxSection. - Refactor BoxSection to use pseudoelements. - Add styles based on Invision
