--- src2/components/BoxAction.scss
@@ -1,37 +1,0 @@
-$box-action-btn-pad-vertical: 16px; 
-$box-action-btn-pad-horizontal: 46px; 
-$box-action-btn-radius: 6px;
-
-.box-action {
-    background-position: center;
-    background-size: cover;
-    background-repeat: no-repeat;
-    
-    &__titles-container {
-        text-align: center;
-        color: white;
-    }
-    
-    // TODO: se va
-    &__button {
-		background-color: red;
-		padding: $box-action-btn-pad-vertical $box-action-btn-pad-horizontal;
-		border: 0;
-		border-radius: $box-action-btn-radius;
-		color: white; 
-    }
-
-    // TODO: se va
-    &__button-wrapper {
-        position: relative;
-    }
-
-    // TODO: se va
-    &__button-color {
-        position: absolute;
-        height: 100%;
-        width: 100%;
-        background: red;
-        z-index: -1;
-    }
-}
