+++ src2/pages/Home.scss
@@ -32,14 +32,6 @@ html, body {
     font-weight: 400;
 }
 
-// h1 {
-//     font-size: 72px;
-// }
-
-// h2 {
-//     font-size: 65px;
-// }
-
 h1 {  
     font-size: $mod_1*$mod_1*$mod_1*$mod_1 *1rem; 
     font-family: $title-font-family;
